Verify id customer

+++ src/main/java/com/br/clean/arch/application/usecases/card/DeleteCard.java
@@ -2,6 +2,7 @@ package com.br.clean.arch.application.usecases.card;
 
 import com.br.clean.arch.application.gateways.card.RepositoryCard;
 import com.br.clean.arch.domain.entitie.card.Card;
+import com.br.clean.arch.domain.entitie.card.exeptions.IncorrectCardExpetion;
 
 public class DeleteCard {
 
@@ -12,6 +13,9 @@ public class DeleteCard {
 	}
 	
 	public Card deleteCard(Long id) {
+		if(repository.findById(id).isEmpty()) {
+			throw new IncorrectCardExpetion("Card not found with id : " + id);
+		}
 		return repository.deleteCard(id);
 	}
 }
